reverted memset

+++ libc/mem.c
@@ -7,8 +7,8 @@ void memory_copy(char *source, char *dest, int nbytes) {
     }
 }
 
-void memory_set(u32 *dest, u8 val, u32 len) {
-    u32 *temp = (u32 *)dest;
+void memory_set(u8 *dest, u8 val, u32 len) {
+    u8 *temp = (u8 *)dest;
     for ( ; len != 0; len--) *temp++ = val;
 }
